Needed to change type to float using osqp eigen

+++ include/FSRModel.h
@@ -11,8 +11,8 @@
 #include <vector>
 
 #include <Eigen/Dense>
-using MatrixXf = Eigen::MatrixXf;
-using VectorXf = Eigen::VectorXf; 
+using MatrixXd = Eigen::MatrixXd;
+using VectorXd = Eigen::VectorXd; 
 /**
  * @brief A Finite Step Response model object. C++ class object holding the FSR model given a spesific format of the step response coefficients.
  * A MPC configuration is also passed as input in order shape the system matrices for the MPC algorithm. 
@@ -26,19 +26,19 @@ private:
     int M_; /** Control horizon */
     int W_; /** Time delay coefficient */
 
-    VectorXf u_; /** Manipulated variables, U(k-N), n_MV */
-    VectorXf y_; /** Controlled variables n_CV */
+    VectorXd u_; /** Manipulated variables, U(k-N), n_MV */
+    VectorXd y_; /** Controlled variables n_CV */
 
-    VectorXf u_K_; /** Denotes U(k-1), n_MV */
-    VectorXf du_tilde_; /** Post change in actuation (n_MV*(N-W-1)) */
+    VectorXd u_K_; /** Denotes U(k-1), n_MV */
+    VectorXd du_tilde_; /** Post change in actuation (n_MV*(N-W-1)) */
 
-    VectorXf** pp_SR_vec_; /** Matrix of Eigen::VectorXf holding every n_CV * n_MV step response */
-    MatrixXf** pp_SR_mat_; /** Tensor of Eigen::MatrixXf representing the SISO prediction (P-W,M) times (n_CV, n_MV) */
-    MatrixXf theta_; /** Matrix of all SISO predictions (n_CV*(P-W), n_MV*M) */
-    MatrixXf phi_; /** Past step coefficients (n_CV*P-W, n_MV*(N-W-1)) */
-    MatrixXf psi_; /** Last step coefficient matrix, (n_CV (P-W), n_MV)*/
+    VectorXd** pp_SR_vec_; /** Matrix of Eigen::VectorXd holding every n_CV * n_MV step response */
+    MatrixXd** pp_SR_mat_; /** Tensor of Eigen::MatrixXd representing the SISO prediction (P-W,M) times (n_CV, n_MV) */
+    MatrixXd theta_; /** Matrix of all SISO predictions (n_CV*(P-W), n_MV*M) */
+    MatrixXd phi_; /** Past step coefficients (n_CV*P-W, n_MV*(N-W-1)) */
+    MatrixXd psi_; /** Last step coefficient matrix, (n_CV (P-W), n_MV)*/
 
-    void AllocateAndDeepCopy(VectorXf** SR);
+    void AllocateAndDeepCopy(VectorXd** SR);
 
     /**
      * @brief Set pp_SR_mat, by generating SISO predictions from pp_SR_vec
@@ -48,9 +48,9 @@ private:
      * @brief Set lower triangular matrix based on the prediction vector
      * 
      * @param pred_vec prediction vector 
-     * @param S Eigen::MatrixXf to be filled with lower triangular SISO predictions
+     * @param S Eigen::MatrixXd to be filled with lower triangular SISO predictions
      */
-    void setLowerTriangularMatrix(const VectorXf& pred_vec, MatrixXf& S);
+    void setLowerTriangularMatrix(const VectorXd& pred_vec, MatrixXd& S);
     
     /**
      * @brief Set the Theta Matrix object. By filling the matrix with SISO preductions from pp_SR_mat
@@ -68,16 +68,16 @@ private:
      * @param vec Vector to be filled n_MV times into phi
      * @param row row indicator
      */
-    void FillRowPhi(const VectorXf& vec, const int& row);
+    void FillRowPhi(const VectorXd& vec, const int& row);
 
     /**
      * @brief Pad the vector, by appending Sn (last element of vec) pad times
      * 
-     * @param vec Eigen::VectorXf to be padded
+     * @param vec Eigen::VectorXd to be padded
      * @param pad Number of pads
-     * @return Eigen::VectorXf Padded vector 
+     * @return Eigen::VectorXd Padded vector 
      */
-    VectorXf PadVec(VectorXf& vec, int pad);
+    VectorXd PadVec(VectorXd& vec, int pad);
 
     void setPsi();
 
@@ -85,14 +85,14 @@ public:
     /**
      * @brief The constructor. Constructing the object allocating memory for the SISO prediction matric
      */
-    FSRModel(VectorXf** SR, int n_CV, int n_MV, int N, int P, int M, int W, 
-            const std::vector<float>& init_u, const std::vector<float>& init_y);
+    FSRModel(VectorXd** SR, int n_CV, int n_MV, int N, int P, int M, int W, 
+            const std::vector<double>& init_u, const std::vector<double>& init_y);
     /**
      * @brief The destructor. Freeing the memory allocated in the constructor
      */
     ~FSRModel();
 
-    void UpdateU(const VectorXf& du, const MatrixXf& du_gamma);
+    void UpdateU(const VectorXd& du, const MatrixXd& du_gamma);
 
     /** Get functions */
     int getP() const { return P_; }
@@ -101,10 +101,10 @@ public:
     int getN_CV() const { return n_CV_; }
     int getN_MV() const { return n_MV_; }
     int getN() const { return N_; }
-    MatrixXf getTheta() const { return theta_; }
-    MatrixXf getPhi() const { return phi_; }
-    VectorXf getUK() const { return u_K_; }
-    VectorXf getLambda() const { return phi_*du_tilde_ + psi_*u_; }
+    MatrixXd getTheta() const { return theta_; }
+    MatrixXd getPhi() const { return phi_; }
+    VectorXd getUK() const { return u_K_; }
+    VectorXd getLambda() const { return phi_*du_tilde_ + psi_*u_; }
 
     /** Print functions */
     void PrintPPSR(int i, int j);
