+++ src/IO/data_objects.cc
@@ -26,14 +26,14 @@ CVData::CVData(const json& cv_data, int n_MV, int n_CV, int N, int T) : n_CV_{n_
         throw std::invalid_argument("n_CV does not coincide with CV");
     }
     y_ref_.resize(T*n_MV);
-    // Allocate matrix of Eigen::VectorXf
-    pp_SR_vec_ = new VectorXf*[n_CV_];
+    // Allocate matrix of Eigen::VectorXd
+    pp_SR_vec_ = new VectorXd*[n_CV_];
     for (int i = 0; i < n_CV_; ++i) {
-        pp_SR_vec_[i] = new VectorXf[n_MV_];
+        pp_SR_vec_[i] = new VectorXd[n_MV_];
     }
     for (int row = 0; row < n_CV_; row++) {
         for (int col = 0; col < n_MV_; col++) {
-            pp_SR_vec_[row][col] = VectorXf::Zero(N);
+            pp_SR_vec_[row][col] = VectorXd::Zero(N);
         }
     }
     for (int outputs = 0; outputs < n_CV; outputs++) {
@@ -67,9 +67,9 @@ CVData& CVData::operator=(const CVData& rhs) {
     y_ref_ = rhs.y_ref_;
 
     // Deep copying
-    pp_SR_vec_ = new VectorXf*[n_CV_];
+    pp_SR_vec_ = new VectorXd*[n_CV_];
     for (int i = 0; i < n_CV_; ++i) {
-        pp_SR_vec_[i] = new VectorXf[n_MV_];
+        pp_SR_vec_[i] = new VectorXd[n_MV_];
     }
     for (int row = 0; row < n_CV_; row++) {
         for (int col = 0; col < n_MV_; col++) {
@@ -79,7 +79,7 @@ CVData& CVData::operator=(const CVData& rhs) {
     return *this;
 }
 
-void FillReference(const json& ref_data, VectorXf& ref, int start_index, int interval) {
+void FillReference(const json& ref_data, VectorXd& ref, int start_index, int interval) {
     for (int i = start_index; i < start_index+interval; i++) {
             ref[i] = ref_data.at(i);
     }  
@@ -88,7 +88,7 @@ void FillReference(const json& ref_data, VectorXf& ref, int start_index, int int
 void CVData::FillSR(const json& s_data) {
     for (int i = 0; i < n_CV_; i++) {
         for (int j = 0; j < n_MV_; j++) {
-            VectorXf vec = VectorXf::Zero(N_);
+            VectorXd vec = VectorXd::Zero(N_);
             for (int k = 0; k < N_; k++) {
                 vec(k) = s_data.at(i).at(k);
             }
@@ -97,7 +97,7 @@ void CVData::FillSR(const json& s_data) {
     }       
 }
 
-VectorXf CVData::getYRef(int P, int k) {
+VectorXd CVData::getYRef(int P, int k) {
     return y_ref_(Eigen::seq(k, P+k));
 }
