+++ src/model/FSRModel.cc
@@ -9,20 +9,20 @@
 #include <vector>
 
 #include <Eigen/Dense>
-using MatrixXf = Eigen::MatrixXf;
-using VectorXf = Eigen::VectorXf; 
+using MatrixXd = Eigen::MatrixXd;
+using VectorXd = Eigen::VectorXd; 
 
-FSRModel::FSRModel(VectorXf** SR, int n_CV, int n_MV, int N, int P, int M, int W,
-                   const std::vector<float>& init_u, const std::vector<float>& init_y) :
+FSRModel::FSRModel(VectorXd** SR, int n_CV, int n_MV, int N, int P, int M, int W,
+                   const std::vector<double>& init_u, const std::vector<double>& init_y) :
                      n_CV_{n_CV}, n_MV_{n_MV}, N_{N}, P_{P}, M_{M}, W_{W} {
-    theta_ = MatrixXf::Zero(n_CV*(P-W), n_MV*M);
+    theta_ = MatrixXd::Zero(n_CV*(P-W), n_MV*M);
     phi_.resize(n_CV*(P-W), n_MV*(N-W-1));
     psi_.resize(n_CV*(P-W), n_MV);
-    du_tilde_ = VectorXf::Zero(n_MV * (N-W-1));
+    du_tilde_ = VectorXd::Zero(n_MV * (N-W-1));
 
-    u_K_ = VectorXf::Map(init_u.data(), init_u.size());
-    u_ = VectorXf::Map(init_u.data(), init_u.size());
-    y_ = VectorXf::Map(init_y.data(), init_y.size());
+    u_K_ = VectorXd::Map(init_u.data(), init_u.size());
+    u_ = VectorXd::Map(init_u.data(), init_u.size());
+    y_ = VectorXd::Map(init_y.data(), init_y.size());
 
     AllocateAndDeepCopy(SR); 
     // Setting matrix member variables
@@ -41,21 +41,21 @@ FSRModel::~FSRModel() {
     delete[] pp_SR_mat_;
 }
 
-void FSRModel::AllocateAndDeepCopy(VectorXf** SR) {
+void FSRModel::AllocateAndDeepCopy(VectorXd** SR) {
     // Allocate memory
-    pp_SR_mat_ = new MatrixXf*[n_CV_];
+    pp_SR_mat_ = new MatrixXd*[n_CV_];
     for (int i = 0; i < n_CV_; ++i) {
-        pp_SR_mat_[i] = new MatrixXf[n_MV_];
+        pp_SR_mat_[i] = new MatrixXd[n_MV_];
     }
     for (int row = 0; row < n_CV_; row++) {
         for (int col = 0; col < n_MV_; col++) {
-            pp_SR_mat_[row][col] = MatrixXf::Zero(P_, M_);
+            pp_SR_mat_[row][col] = MatrixXd::Zero(P_, M_);
         }
     }
     // Deep copy SR into pp_SR_vec
-    pp_SR_vec_ = new VectorXf*[n_CV_];
+    pp_SR_vec_ = new VectorXd*[n_CV_];
     for (int i = 0; i < n_CV_; ++i) {
-        pp_SR_vec_[i] = new VectorXf[n_MV_];
+        pp_SR_vec_[i] = new VectorXd[n_MV_];
     }
     for (int row = 0; row < n_CV_; row++) {
         for (int col = 0; col < n_MV_; col++) {
@@ -64,7 +64,7 @@ void FSRModel::AllocateAndDeepCopy(VectorXf** SR) {
     }
 }
 
-void FSRModel::setLowerTriangularMatrix(const VectorXf& pred_vec, MatrixXf& S) {
+void FSRModel::setLowerTriangularMatrix(const VectorXd& pred_vec, MatrixXd& S) {
     int n = 0; 
     for (int i = 0; i < M_; i++) {
         for (int j = 0; j < P_-n; j++) {
@@ -76,7 +76,7 @@ void FSRModel::setLowerTriangularMatrix(const VectorXf& pred_vec, MatrixXf& S) {
 
 void FSRModel::setSRMatrix() {
     for (int i = 0; i < n_CV_; i++) {
-        MatrixXf S = MatrixXf::Zero(P_, M_);
+        MatrixXd S = MatrixXd::Zero(P_, M_);
         for (int j = 0; j < n_MV_; j++) {
             setLowerTriangularMatrix(pp_SR_vec_[i][j], S);
             pp_SR_mat_[i][j] = S(Eigen::seq(W_, Eigen::last), Eigen::seq(0, Eigen::last));
@@ -96,26 +96,26 @@ void FSRModel::setPhiMatrix() {
     // NB: Need supervision for padding Sn 
     for (int i = 0; i < n_CV_; i++) {
         for (int j = 0; j < n_MV_; j++) { 
-            VectorXf vec = pp_SR_vec_[i][j](Eigen::seq(P_-W_+j,Eigen::last));
+            VectorXd vec = pp_SR_vec_[i][j](Eigen::seq(P_-W_+j,Eigen::last));
             if (vec.size() != N_-P_) { // Pad vector
                 int pad = vec.size() - (N_-W_-1);
-                VectorXf vec = PadVec(vec, pad);
+                VectorXd vec = PadVec(vec, pad);
             }
             FillRowPhi(vec, i);
         }
     }
 }
 
-VectorXf FSRModel::PadVec(VectorXf& vec, int pad) {
-    float num = vec(Eigen::last); // Accessing Sn
-    VectorXf num_vec = VectorXf::Constant(pad, num);
-    VectorXf padded_vec(vec.size() + num_vec.size());
-    padded_vec << vec; // Concatinating two Eigen::VectorXf
+VectorXd FSRModel::PadVec(VectorXd& vec, int pad) {
+    double num = vec(Eigen::last); // Accessing Sn
+    VectorXd num_vec = VectorXd::Constant(pad, num);
+    VectorXd padded_vec(vec.size() + num_vec.size());
+    padded_vec << vec; // Concatinating two Eigen::VectorXd
     padded_vec << num_vec;
     return padded_vec; 
 }
 
-void FSRModel::FillRowPhi(const VectorXf& pad_vec, const int& row) {
+void FSRModel::FillRowPhi(const VectorXd& pad_vec, const int& row) {
     for (int i = 0; i < n_MV_; i++) { // Might need to be tested
         for (int j = row; j < P_; j++) {
             phi_(j, Eigen::seq(i*(N_-P_-1), (i+1)*(N_-P_-1))) = pad_vec(Eigen::seq(0, Eigen::last));
@@ -126,13 +126,13 @@ void FSRModel::FillRowPhi(const VectorXf& pad_vec, const int& row) {
 void FSRModel::setPsi() {
     for (int i = 0; i < n_CV_; i++) {
         for (int j = 0; j < n_MV_; j++) {
-            VectorXf vec = VectorXf::Constant(P_-W_, pp_SR_vec_[i][j](Eigen::last));
+            VectorXd vec = VectorXd::Constant(P_-W_, pp_SR_vec_[i][j](Eigen::last));
             psi_.block(i*(P_-W_), j, P_-W_, 1) = vec;
         }
     }
 }
 
-void FSRModel::UpdateU(const VectorXf& du, const MatrixXf& du_gamma) { //du_gamma = Gamma * du 
+void FSRModel::UpdateU(const VectorXd& du, const MatrixXd& du_gamma) { //du_gamma = Gamma * du 
     u_K_ += du; // Update using first col
     
     // Update du_tilde by left shift, adding the optimized du
